pyproject.toml
[tool.poetry]
name = "internetgpt"
version = "0.1.0"
description = ""
authors = ["Rojer Du <rojerdu.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.284"
mypy = "^1.5.0"
langchain = "^0.0.265"
streamlit = "^1.25.0"
openai = "^0.27.8"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
